UWP serialize data Json
