+++ VersionTracking.nuspec
@@ -2,7 +2,7 @@
 <package xmlns="http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd">
 	<metadata minClientVersion="2.8.3">
 		<id>Plugin.VersionTracking</id>
-		<version>1.0.1</version>
+		<version>1.0.2</version>
 		<title>VersionTracking Plugin for Xamarin and Windows</title>
 		<authors>Colby Williams</authors>
 		<owners>Colby Williams</owners>
@@ -34,30 +34,6 @@
 		<file src="VersionTracking\Plugin.VersionTracking.Abstractions\bin\Release\Plugin.VersionTracking.Abstractions.dll" target="lib\dotnet\Plugin.VersionTracking.Abstractions.dll" />
 		<file src="VersionTracking\Plugin.VersionTracking.Abstractions\bin\Release\Plugin.VersionTracking.Abstractions.xml" target="lib\dotnet\Plugin.VersionTracking.Abstractions.xml" />
 		<file src="VersionTracking\Plugin.VersionTracking.Abstractions\bin\Release\Plugin.VersionTracking.Abstractions.pdb" target="lib\dotnet\Plugin.VersionTracking.Abstractions.pdb" />
-
-		<!--Win Phone Silverlight-->
-		<file src="VersionTracking\Plugin.VersionTracking.WindowsPhone8\bin\Release\Plugin.VersionTracking.dll" target="lib\wp8\Plugin.VersionTracking.dll" />
-		<file src="VersionTracking\Plugin.VersionTracking.WindowsPhone8\bin\Release\Plugin.VersionTracking.xml" target="lib\wp8\Plugin.VersionTracking.xml" />
-		<file src="VersionTracking\Plugin.VersionTracking.WindowsPhone8\bin\Release\Plugin.VersionTracking.pdb" target="lib\wp8\Plugin.VersionTracking.pdb" />
-		<file src="VersionTracking\Plugin.VersionTracking.Abstractions\bin\Release\Plugin.VersionTracking.Abstractions.dll" target="lib\wp8\Plugin.VersionTracking.Abstractions.dll" />
-		<file src="VersionTracking\Plugin.VersionTracking.Abstractions\bin\Release\Plugin.VersionTracking.Abstractions.xml" target="lib\wp8\Plugin.VersionTracking.Abstractions.xml" />
-		<file src="VersionTracking\Plugin.VersionTracking.Abstractions\bin\Release\Plugin.VersionTracking.Abstractions.pdb" target="lib\wp8\Plugin.VersionTracking.Abstractions.pdb" />
-
-		<!--Win Phone 81-->
-		<file src="VersionTracking\Plugin.VersionTracking.WindowsPhone81\bin\Release\Plugin.VersionTracking.dll" target="lib\wpa81\Plugin.VersionTracking.dll" />
-		<file src="VersionTracking\Plugin.VersionTracking.WindowsPhone81\bin\Release\Plugin.VersionTracking.xml" target="lib\wpa81\Plugin.VersionTracking.xml" />
-		<file src="VersionTracking\Plugin.VersionTracking.WindowsPhone81\bin\Release\Plugin.VersionTracking.pdb" target="lib\wpa81\Plugin.VersionTracking.pdb" />
-		<file src="VersionTracking\Plugin.VersionTracking.Abstractions\bin\Release\Plugin.VersionTracking.Abstractions.dll" target="lib\wpa81\Plugin.VersionTracking.Abstractions.dll" />
-		<file src="VersionTracking\Plugin.VersionTracking.Abstractions\bin\Release\Plugin.VersionTracking.Abstractions.xml" target="lib\wpa81\Plugin.VersionTracking.Abstractions.xml" />
-		<file src="VersionTracking\Plugin.VersionTracking.Abstractions\bin\Release\Plugin.VersionTracking.Abstractions.pdb" target="lib\wpa81\Plugin.VersionTracking.Abstractions.pdb" />
-
-		<!--WinStore-->
-		<file src="VersionTracking\Plugin.VersionTracking.WindowsStore\bin\Release\Plugin.VersionTracking.dll" target="lib\win8\Plugin.VersionTracking.dll" />
-		<file src="VersionTracking\Plugin.VersionTracking.WindowsStore\bin\Release\Plugin.VersionTracking.xml" target="lib\win8\Plugin.VersionTracking.xml" />
-		<file src="VersionTracking\Plugin.VersionTracking.WindowsStore\bin\Release\Plugin.VersionTracking.pdb" target="lib\win8\Plugin.VersionTracking.pdb" />
-		<file src="VersionTracking\Plugin.VersionTracking.Abstractions\bin\Release\Plugin.VersionTracking.Abstractions.dll" target="lib\win8\Plugin.VersionTracking.Abstractions.dll" />
-		<file src="VersionTracking\Plugin.VersionTracking.Abstractions\bin\Release\Plugin.VersionTracking.Abstractions.xml" target="lib\win8\Plugin.VersionTracking.Abstractions.xml" />
-		<file src="VersionTracking\Plugin.VersionTracking.Abstractions\bin\Release\Plugin.VersionTracking.Abstractions.pdb" target="lib\win8\Plugin.VersionTracking.Abstractions.pdb" />
 
 		<!--Xamarin.Android-->
 		<file src="VersionTracking\Plugin.VersionTracking.Android\bin\Release\Plugin.VersionTracking.dll" target="lib\MonoAndroid10\Plugin.VersionTracking.dll" />
