+++ VersionTracking/Plugin.VersionTracking.Android/VersionTrackingImplementation.cs
@@ -258,7 +258,10 @@ namespace Plugin.VersionTracking
             if (FirstLaunchForBuild(build)) block?.Invoke();
         }
 
-
+        /// <summary>
+        /// Return all values into a string
+        /// </summary>
+        /// <returns>All values</returns>
         public override string ToString()
         {
             var sb = new StringBuilder();
