+++ VersionTracking/Plugin.VersionTracking.UWP/Plugin.VersionTracking.UWP.csproj
@@ -11,7 +11,7 @@
     <AssemblyName>Plugin.VersionTracking</AssemblyName>
     <DefaultLanguage>en-US</DefaultLanguage>
     <TargetPlatformIdentifier>UAP</TargetPlatformIdentifier>
-    <TargetPlatformVersion>10.0.10586.0</TargetPlatformVersion>
+    <TargetPlatformVersion>10.0.14393.0</TargetPlatformVersion>
     <TargetPlatformMinVersion>10.0.10240.0</TargetPlatformMinVersion>
     <MinimumVisualStudioVersion>14</MinimumVisualStudioVersion>
     <FileAlignment>512</FileAlignment>
@@ -109,13 +109,11 @@
     <None Include="project.json" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="..\Plugin.VersionTracking.WindowsPhone81\VersionTrackingImplementation.cs">
-      <Link>VersionTrackingImplementation.cs</Link>
-    </Compile>
     <Compile Include="..\Plugin.VersionTracking\CrossVersionTracking.cs">
       <Link>CrossVersionTracking.cs</Link>
     </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="VersionTrackingImplementation.cs" />
     <EmbeddedResource Include="Properties\Plugin.VersionTracking.UWP.rd.xml" />
   </ItemGroup>
   <ItemGroup>
