+++ VersionTracking/Plugin.VersionTracking.UWP/project.json
@@ -1,6 +1,7 @@
 ﻿{
   "dependencies": {
-    "Microsoft.NETCore.UniversalWindowsPlatform": "5.0.0"
+    "Microsoft.NETCore.UniversalWindowsPlatform": "5.3.3",
+    "Newtonsoft.Json": "9.0.1"
   },
   "frameworks": {
     "uap10.0": {}
